+++ src/main/java/com/podzilla/order/model/OrderLocation.java
@@ -1,13 +1,10 @@
 package com.podzilla.order.model;
 
-import jakarta.persistence.Entity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.Data;
-
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -16,6 +13,6 @@ import java.math.BigDecimal;
 @Data
 public class OrderLocation {
 
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private double latitude;
+    private double longitude;
 }
